+++ ui/tucu-ui/src/datatable/components/DataTable/index.ts
@@ -1,0 +1,3 @@
+export { default as DataTableComponent } from './DataTableComponent';
+export { default as DataTable } from './DataTable';
+export { default as ColumnsVisibilityManager } from './ColumnsVisibilityManager';